Rename CSharpVBBindingConfigProvider to RoslynBindingConfigProvider
