+++ src/ConnectedMode.UnitTests/Binding/CompositeBindingConfigProviderTests.cs
@@ -48,7 +48,7 @@ namespace SonarLint.VisualStudio.ConnectedMode.UnitTests.Binding
 
             // Assert
             testSubject.Providers.Count().Should().Be(2);
-            testSubject.Providers.Select(x => x.GetType()).Should().BeEquivalentTo(typeof(NonRoslynDummyBindingConfigProvider), typeof(CSharpVBBindingConfigProvider));
+            testSubject.Providers.Select(x => x.GetType()).Should().BeEquivalentTo(typeof(NonRoslynDummyBindingConfigProvider), typeof(RoslynBindingConfigProvider));
         }
 
         [TestMethod]
