+++ src/ConnectedMode/Binding/CompositeBindingConfigProvider.cs
@@ -38,7 +38,7 @@ namespace SonarLint.VisualStudio.ConnectedMode.Binding
 
         [ImportingConstructor]
         public CompositeBindingConfigProvider(ISonarQubeService sonarQubeService, ILogger logger, ILanguageProvider languageProvider)
-            : this(new NonRoslynDummyBindingConfigProvider(languageProvider), new CSharpVBBindingConfigProvider(sonarQubeService, logger, languageProvider))
+            : this(new NonRoslynDummyBindingConfigProvider(languageProvider), new RoslynBindingConfigProvider(sonarQubeService, logger, languageProvider))
         {
         }
